requirements.txt
python-telegram-bot==20.3
httpx==0.24.1
